Nu fungerar det att hämta alla assignments från databasen. Det var köttigare än det borde.. Assignments hamnar på väldigt konstiga platser. Varför?

+++ klient-projekttermin/src/com/example/klien_projekttermin/databaseProvider/Database.java
@@ -115,8 +115,8 @@ public class Database {
 					ass.getAssignmentStatus().toString());
 			values.put(AssignmentTable.COLUMN_CAMERAIMAGE,
 					bitmapToByteArray(ass.getCameraImage()));
-			values.put(AssignmentTable.COLUMN_LAT, Long.toString(ass.getLat()));
-			values.put(AssignmentTable.COLUMN_LON, Long.toString(ass.getLon()));
+			values.put(AssignmentTable.COLUMN_LAT, Double.toString(ass.getLat()));
+			values.put(AssignmentTable.COLUMN_LON, Double.toString(ass.getLon()));
 			values.put(AssignmentTable.COLUMN_NAME, ass.getName());
 			values.put(AssignmentTable.COLUMN_SENDER, ass.getSender());
 			values.put(AssignmentTable.COLUMN_AGENTS, agents);
@@ -337,8 +337,8 @@ public class Database {
 					ass.getAssignmentStatus().toString());
 			values.put(AssignmentTable.COLUMN_CAMERAIMAGE,
 					bitmapToByteArray(ass.getCameraImage()));
-			values.put(AssignmentTable.COLUMN_LAT, Long.toString(ass.getLat()));
-			values.put(AssignmentTable.COLUMN_LON, Long.toString(ass.getLon()));
+			values.put(AssignmentTable.COLUMN_LAT, Double.toString(ass.getLat()));
+			values.put(AssignmentTable.COLUMN_LON, Double.toString(ass.getLon()));
 			values.put(AssignmentTable.COLUMN_NAME, ass.getName());
 			values.put(AssignmentTable.COLUMN_SENDER, ass.getSender());
 			values.put(AssignmentTable.COLUMN_SITENAME, ass.getSiteName());
