+++ klient-projekttermin/src/models/Assignment.java
@@ -18,9 +18,9 @@ public class Assignment implements ModelInterface {
 	// Namnet på uppdraget
 	private String name;
 	// Latitud för uppdragspositionen
-	private long lat;
+	private double lat;
 	// Longitud för uppdragspositionen
-	private long lon;
+	private double lon;
 	// JSON-sträng med WSG-punkter för polygonmarkering av region på kartan
 	private String region;
 	// Användarnamnet på den person som skapade uppdraget.
@@ -153,7 +153,7 @@ public class Assignment implements ModelInterface {
 	 * @param streetName
 	 * @param siteName
 	 */
-	public Assignment(String name, long lat, long lon, 
+	public Assignment(String name, double lat, double lon, 
 			String sender, boolean externalMission,
 			String assignmentDescription, String timeSpan,
 			AssignmentStatus assignmentStatus, String streetName,
@@ -186,7 +186,7 @@ public class Assignment implements ModelInterface {
 	 * @param streetName
 	 * @param siteName
 	 */
-	public Assignment(String name, long lat, long lon, 
+	public Assignment(String name, double lat, double lon, 
 			String sender, boolean externalMission,
 			String assignmentDescription, String timeSpan,
 			AssignmentStatus assignmentStatus, Bitmap cameraImage,
@@ -285,7 +285,7 @@ public class Assignment implements ModelInterface {
 	 * @param streetName
 	 * @param siteName
 	 */
-	public Assignment(long id, String name, long lat, long lon, String region,
+	public Assignment(long id, String name, double lat, double lon, String region,
 			List<Contact> agents,
 			String sender, boolean externalMission,
 			String assignmentDescription, String timeSpan,
@@ -332,11 +332,11 @@ public class Assignment implements ModelInterface {
 		return siteName;
 	}
 
-	public long getLat() {
+	public double getLat() {
 		return lat;
 	}
 
-	public long getLon() {
+	public double getLon() {
 		return lon;
 	}
